requirements.txt
openai
pdfplumber
Pillow 